fix: import CryptoPunks section Framer Motion's components on client-side
